+++ requirements.txt
@@ -9,3 +9,4 @@ google-cloud-texttospeech
 tantivy
 pymongo
 backoff
+deepmultilingualpunctuation
